consumers/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.9
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
python-dotenv>=1.0.0
